fixed small remaining bugs - fixed size of viewpager tabs - inconsistency of loading fragments
