+++ app/src/main/java/ch/hsrmarket/android/CategoryFragment.java
@@ -21,8 +21,9 @@ import ch.hsrmarket.android.model.Other;
 
 public class CategoryFragment extends Fragment implements ApiClient.OnResponseListener{
 
-    private CategoryAdapter adapter;
+    private CategoryAdapter adapter ;
     private Article.Type appointedCategory;
+    private RecyclerView recyclerView;
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
@@ -31,12 +32,9 @@ public class CategoryFragment extends Fragment implements ApiClient.OnResponseLi
         Bundle bundle = getArguments();
         appointedCategory = (Article.Type) bundle.getSerializable("appointedCategory");
 
-        RecyclerView recyclerView = (RecyclerView) root.findViewById(R.id.listCategory);
+        recyclerView = (RecyclerView) root.findViewById(R.id.listCategory);
         RecyclerView.LayoutManager mLayoutManager = new LinearLayoutManager(getActivity().getApplicationContext());
         recyclerView.setLayoutManager(mLayoutManager);
-
-        adapter = new CategoryAdapter<>();
-        recyclerView.setAdapter(adapter);
 
         ApiClient apiClient = new ApiClient();
         apiClient.setOnResponseListener(this);
@@ -65,19 +63,27 @@ public class CategoryFragment extends Fragment implements ApiClient.OnResponseLi
         switch (appointedCategory){
             case BOOK:
                 List<Book> books = (List<Book>) data;
-                adapter.setData(books);
+                adapter = new CategoryAdapter<>(books);
+                recyclerView.setAdapter(adapter);
+
                 break;
             case ELECTRONIC_DEVICE:
                 List<ElectronicDevice> electronicDevices = (List<ElectronicDevice>) data;
-                adapter.setData(electronicDevices);
+                adapter = new CategoryAdapter<>(electronicDevices);
+                recyclerView.setAdapter(adapter);
+
                 break;
             case OFFICE_SUPPLY:
                 List<OfficeSupply> officeSupplies = (List<OfficeSupply>) data;
-                adapter.setData(officeSupplies);
+                adapter = new CategoryAdapter<>(officeSupplies);
+                recyclerView.setAdapter(adapter);
+
                 break;
             case OTHER:
                 List<Other> others = (List<Other>) data;
-                adapter.setData(others);
+                adapter = new CategoryAdapter<>(others);
+                recyclerView.setAdapter(adapter);
+
                 break;
         }
     }
