+++ app/src/main/res/layout/activity_main.xml
@@ -19,7 +19,10 @@
         <android.support.design.widget.TabLayout
             android:id="@+id/tabs"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content" />
+            android:layout_height="wrap_content"
+            app:tabMode="scrollable"
+
+            />
 
     </android.support.design.widget.AppBarLayout>
